+++ tests/Doctrine/Tests/ODM/PHPCR/Mapping/Model/ChildMappingObject.php
@@ -5,18 +5,18 @@ namespace Doctrine\Tests\ODM\PHPCR\Mapping\Model;
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCRODM;
 
 /**
- * A class that contains mapped children via properties 
- * 
+ * A class that contains mapped children via properties
+ *
  * @PHPCRODM\Document
  */
 class ChildMappingObject
 {
     /** @PHPCRODM\Id */
     public $id;
-    
-    /** @PHPCRODM\Child(name="first") */
+
+    /** @PHPCRODM\Child(nodeName="first") */
     public $child1;
-    
-    /** @PHPCRODM\Child(name="second") */
+
+    /** @PHPCRODM\Child(nodeName="second") */
     public $child2;
 }